% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convertToLlama}
\alias{convertToLlama}
\title{Convert an ASScenario scenario object to a llama data object.}
\usage{
convertToLlama(asscenario, measure, feature.steps)
}
\arguments{
\item{asscenario}{[\code{\link{ASScenario}}]\cr
Algorithm selection scenario.}

\item{measure}{[\code{character(1)}]\cr
Measure to use for modeling.
Default is first measure in scenario.}

\item{feature.steps}{[\code{character}]\cr
Which feature steps are allowed?
Default are the default feature steps or all steps
in case no defaults were defined.}
}
\value{
Result of calling \code{\link[llama]{input}}.
}
\description{
For features, mean values are computed across repetitions.
For algorithms, repetitions are not supported at the moment and will result in an error.
}

