% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToLlama.R
\name{fixFeckingPresolve}
\alias{fixFeckingPresolve}
\title{Bakes presolving stuff into a LLAMA data frame.}
\usage{
fixFeckingPresolve(asscenario, ldf)
}
\arguments{
\item{asscenario}{[\code{\link{ASScenario}}]\cr
Algorithm selection scenario.}

\item{ldf}{[\code{LLAMA data frame}]\cr
LLAMA data frame to modify.}
}
\value{
The LLAMA data frame with presolving baked into the algorithm
       performances.
}
\description{
Determines whether any of the feature groups in the LLAMA data frame
presolve any of the instances. If so, the performances of all algorithms
in the portfolio are set to the runtime of the first used feature group
that presolves the respective instance. Furthermore, the success of all
algorithms on those instances is set to true.
}
\details{
These modifications are done on the main LLAMA data and on any test splits.
They are *not* done on the training data. This function should only ever be
used to evaluate the performance of an actual selector that uses features
(i.e. not VBS or single best). Using it in polite company is to be avoided.
}

