% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCosealASScenario.R
\name{getCosealASScenario}
\alias{getCosealASScenario}
\title{Retrieves a scenario from the Coseal Github repository and parses into an S3 object.}
\usage{
getCosealASScenario(name)
}
\arguments{
\item{name}{[\code{character(1)}]\cr
Name of benchmark data set.}
}
\value{
[\code{\link{ASScenario}}]. Description object.
}
\description{
Uses subversion export to retrieve a specific scenario from the official
Coseal Github repository. The scenario is checked out into a temporary directory
and parsed with \code{parseASScenario}.
}

