% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFeatureStepNames.R
\name{getFeatureStepNames}
\alias{getFeatureStepNames}
\title{Returns feature step names of scenario.}
\usage{
getFeatureStepNames(asscenario)
}
\arguments{
\item{asscenario}{[\code{\link{ASScenario}}]\cr
Algorithm selection scenario.}
}
\value{
[\code{character}].
}
\description{
Returns feature step names of scenario.
}

