% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCV.R
\name{getNumberOfCVFolds}
\alias{getNumberOfCVFolds}
\title{Returns number of CV folds.}
\usage{
getNumberOfCVFolds(asscenario)
}
\arguments{
\item{asscenario}{[\code{\link{ASScenario}}]\cr
Algorithm selection scenario.}
}
\value{
[\code{integer(1)}].
}
\description{
Returns number of CV folds.
}

