% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getProvidedFeatures}
\alias{getProvidedFeatures}
\title{Return features that are useable for a given set of feature steps.}
\usage{
getProvidedFeatures(asscenario, steps)
}
\arguments{
\item{asscenario}{[\code{\link{ASScenario}}]\cr
Algorithm selection scenario.}

\item{steps}{[\code{character}]\cr
Feature steps.
Default are all feature steps.}
}
\value{
[\code{character}].
}
\description{
Return features that are useable for a given set of feature steps.
}

