% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarizeFeatureValues.R
\name{summarizeFeatureValues}
\alias{summarizeFeatureValues}
\title{Creates summary data.frame for feature values across all instances.}
\usage{
summarizeFeatureValues(asscenario)
}
\arguments{
\item{asscenario}{[\code{\link{ASScenario}}]\cr
Algorithm selection scenario.}
}
\value{
[\code{data.frame}].
}
\description{
Creates summary data.frame for feature values across all instances.
}

