% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeASScenario.R
\name{writeASScenario}
\alias{writeASScenario}
\title{Writes an algorithm selection scenario to a directory.}
\usage{
writeASScenario(asscenario, path = asscenario$desc$scenario_id)
}
\arguments{
\item{asscenario}{[\code{\link{ASScenario}}]\cr
Algorithm selection scenario.}

\item{path}{[\code{character(1)}]\cr
Path to write scenario to. Default is the name of the scenario.}
}
\description{
Splits an algorithm selection scenario into description, feature
values / runstatus / costs, algorithm performance and cv splits and
saves those data sets as single ARFF files in the given directory.
}
\seealso{
\code{\link{parseASScenario}}
}

