% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatureStepNames.R
\name{getFeatureStepNames}
\alias{getFeatureStepNames}
\title{Returns feature step names of scenario.}
\usage{
getFeatureStepNames(asscenario, type)
}
\arguments{
\item{asscenario}{[\code{\link{ASScenario}}]\cr
Algorithm selection scenario.}

\item{type}{[\code{character(1)}]
Feature type (instance or algorithmic).}
}
\value{
[\code{character}].
}
\description{
Returns feature step names of scenario.
}
