% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getInstanceNames.R
\name{getInstanceNames}
\alias{getInstanceNames}
\title{Returns instance names of scenario.}
\usage{
getInstanceNames(asscenario)
}
\arguments{
\item{asscenario}{[\code{\link{ASScenario}}]\cr
Algorithm selection scenario.}
}
\value{
[\code{character}].
}
\description{
Returns instance names of scenario.
}

