% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarizeAlgoPerf.R
\name{summarizeAlgoPerf}
\alias{summarizeAlgoPerf}
\title{Creates summary data.frame for algorithm performance values across all instances.}
\usage{
summarizeAlgoPerf(asscenario, measure)
}
\arguments{
\item{asscenario}{[\code{\link{ASScenario}}]\cr
Algorithm selection scenario.}

\item{measure}{[\code{character(1)}]\cr
Selected measure.
Default is first measure in scenario.}
}
\value{
[\code{data.frame}].
}
\description{
Creates summary data.frame for algorithm performance values across all instances.
}

