% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeFeatureValues.R
\name{summarizeFeatureValues}
\alias{summarizeFeatureValues}
\title{Creates summary data.frame for feature values across all instances.}
\usage{
summarizeFeatureValues(asscenario, type)
}
\arguments{
\item{asscenario}{[\code{\link{ASScenario}}]\cr
Algorithm selection scenario.}

\item{type}{[\code{character(1)}]\cr
Feature type (instance or algorithmic).}
}
\value{
[\code{data.frame}].
}
\description{
Creates summary data.frame for feature values across all instances.
}
